requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
python-dotenv>=1.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
sqlalchemy>=2.0.0
pydantic>=2.8.0
email-validator>=2.2.0
bcrypt>=4.0.0
tzdata>=2024.1  # Windows needs this for ZoneInfo; no-op on macOS/Linux
